Reorganized the directory for better navigation
