+++ Projects and Applications/Quiz App in Python/index.py
