+++ Python/Reverse Stack.py
