+++ Python/pattern.py
