+++ Python/prime.py
