+++ Python/star.py
